Color is now stored for each pixel for every path and summed up at the end to display the final color

+++ main/java/app/android/pathtrace/pathtracer/RenderObject.kt
@@ -16,5 +16,5 @@ abstract class RenderObject(var col : Col, var isLight : Boolean) {
         return this.normal
     }
 
-    abstract fun intersect(r : Ray) : Double
+    abstract fun intersect(r : Ray) : Float
 }
